Change compilation pass
The default level-2 optimization expands the classically-controlled CZ gates suboptimally. The level-3 optimization does it correctly, but also removes barriers (which we want to keep). The compromise is to take the level-3 optimization and filter out the barrier-removing pass (and also the Pauli simplification pass, which does not work in the presence of barriers).