Ansatz based on "diamond" contractions
The motivation for this implementation is the discovery that, at least on my machine, applying a single 1-, 2-, 3-, or 4-qubit unitary to a 12-qubit state takes about the same amount of time. This suggests that we can optimize the state computation by first combining blocks of 1- and 2-qubit gates in a "diamond" pattern into 4-qubit unitaries before multiplying them against the state vector.

I doubt that I will use this implementation going forward, as the performance boost may well be machine-specific rather than a generic improvement. This is just to keep a record of its existence.